remoção dos metodos referentes ao usuario da entidade empresa, adicionados ao seu controlador
